requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow 
apache-airflow-providers-apache-spark
pyspark
pymysql 
apache-airflow-providers-mysql 
boto3
pandas
gdown